requirements.txt
praw==7.7.1
kafka-python==2.0.2
python-dotenv==1.0.0
snowflake-connector-python==3.0.3
pandas==2.0.3
textblob==0.17.1
requests==2.31.0
pytz==2023.3
rake-nltk==1.0.6
nltk==3.8.1
